Removed item: "Optional commit API change to branch..."
In the future we will move to a model where changes to APIs are published to source control systems (and perhaps other places) on demand rather than source control being the primary storage for the API content.  At that time we can work on the flow (pull request, branching, etc).